backend/requirements.txt
asgiref
Django
django-cors-headers
django-rest-framework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
